requirements.txt
bidict==0.23.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.3
colorama==0.4.6
cs50==9.3.4
dlib @ file:./dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dnspython==2.6.1
email_validator==2.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
flask-mailman==1.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
passlib==1.7.4
pillow==10.3.0
pluggy==1.5.0
pytest==8.2.1
pytest-flask==1.3.0
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.2
schedule==1.2.2
simple-websocket==1.0.0
SQLAlchemy==2.0.30
SQLAlchemy-serializer==1.4.12
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
